Add listToken name to <List> component, add an alert for logging out to the logOut function

+++ src/App.jsx
@@ -61,9 +61,14 @@ export function App() {
 	}, [listToken]);
 
 	function logOut() {
-		localStorage.clear();
-		setListToken('');
-		navigateTo('/');
+		const confirm = window.confirm(
+			`If you're ready to log out, make sure to write down your list name before you click ok! It is ${listToken}`,
+		);
+		if (confirm) {
+			localStorage.clear();
+			setListToken('');
+			navigateTo('/');
+		}
 	}
 
 	return (
